feat: Simulator 컴포넌트화 -> create(방 새로 만들기) : edit(내방 수정) 모드 분리
